Add :apply_transformations option to gradient methods
As discussed on #894, we don't want to break existing code that uses Prawn
where the coder has manually calculated and applied the transformation to
the gradient co-ordinate arguments.  We add an :apply_transformations option
that defaults to false and a warning when the option is not supplied and a
transformation has been carried out.  It's expected this will default to
true in Prawn v3.